functions/requirements.txt
fitdecode==0.7.0
google-cloud-bigquery==2.7.*
google-cloud-firestore=2.0.*
google-cloud-pubsub==2.3.*
google-cloud-storage==1.36.*
pyyaml==5.3.*
requests==2.22.*
